offst-common: Refactored ser_utils into separate files.
