+++ components/common/src/ser_utils/mod.rs
@@ -1,0 +1,11 @@
+pub mod ser_b64;
+pub mod ser_map_b64_any;
+pub mod ser_map_str_any;
+pub mod ser_map_str_str;
+pub mod ser_option_b64;
+pub mod ser_seq_b64;
+pub mod ser_string;
+pub mod ser_vec_b64;
+
+#[cfg(test)]
+mod tests;
